requirements.txt
gradio
tensorflow
numpy
pillow
numpy
random
shutil